Создан ForecastPresenter. Обновлена логика Injector.

+++ app/src/main/java/com/yamblz/voltek/weather/data/api/weather/Coord.java
@@ -5,8 +5,8 @@ import com.squareup.moshi.Json;
 public class Coord {
 
     @Json(name = "lon")
-    public Integer lon;
+    public Double lon;
 
     @Json(name = "lat")
-    public Integer lat;
+    public Double lat;
 }
